Add files via upload
Unfortunately the folder has gotten too big for a direct upload, but this zip file contains my new code that allows the user to move a more realistic ocean with foam and waves, as well as switch between different ocean color palettes. The ocean is relatively lightweight, but metaXR does disable it by turning off opaque textures in the URP pipeline asset, so I added a script that looks for the specific pipeline asset in this zipped folder and automatically change it to accomodate the ocean shader graphed materials. Still a work in progress!